Fashion_blog_API/src/main/java/com/productblog/controllers/UserController.java
package com.productblog.controllers;


import com.productblog.dtos.UserDto;
import com.productblog.services.UserService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

import java.util.List;

@RestController
@RequestMapping(value = "/blog/api/user")
public class UserController {

    private final UserService userService;

    @Autowired
    public UserController(UserService userService) {
        this.userService = userService;
    }

    @PostMapping("/create")
    public ResponseEntity<UserDto> createUser(@RequestBody UserDto userDto){
       return userService.createUser(userDto);
    }

    @PutMapping("/update/{id}")
    public ResponseEntity<UserDto> updateUser(@PathVariable long id, @RequestBody UserDto userDto){
        return userService.updateUser(userDto, id);
    }

    @GetMapping("/{userId}")
    public ResponseEntity<UserDto> findUserById(@PathVariable("userId") long id){
        return userService.fetchUserById(id);
    }

    @GetMapping("/all")
    public ResponseEntity<List<UserDto>> findUsers(){
        return userService.fetchUsers();
    }


}
